+++ sympy/concrete/products.py
@@ -1,15 +1,19 @@
 from typing import Tuple as tTuple
 
+from .expr_with_intlimits import ExprWithIntLimits
+from .summations import Sum, summation, _dummy_with_inherited_properties_concrete
 from sympy.core.expr import Expr
+from sympy.core.exprtools import factor_terms
+from sympy.core.function import Derivative
 from sympy.core.mul import Mul
 from sympy.core.singleton import S
-from sympy.concrete.expr_with_intlimits import ExprWithIntLimits
-from sympy.core.exprtools import factor_terms
-from sympy.functions.elementary.exponential import exp, log
-from sympy.polys import quo, roots
-from sympy.simplify import powsimp
-from sympy.core.function import Derivative
 from sympy.core.symbol import Dummy, Symbol
+from sympy.functions.combinatorial.factorials import RisingFactorial
+from sympy.functions.elementary.exponential import exp, log
+from sympy.functions.special.tensor_functions import KroneckerDelta
+from sympy.polys import quo, roots
+from sympy.simplify.powsimp import powsimp
+from sympy.simplify.simplify import product_simplify
 
 
 class Product(ExprWithIntLimits):
@@ -196,7 +200,6 @@ class Product(ExprWithIntLimits):
         return obj
 
     def _eval_rewrite_as_Sum(self, *args, **kwargs):
-        from sympy.concrete.summations import Sum
         return exp(Sum(log(self.function), *self.limits))
 
     @property
@@ -252,8 +255,6 @@ class Product(ExprWithIntLimits):
         # variables with matching assumptions
         reps = {}
         for xab in self.limits:
-            # Must be imported here to avoid circular imports
-            from .summations import _dummy_with_inherited_properties_concrete
             d = _dummy_with_inherited_properties_concrete(xab)
             if d:
                 reps[xab[0]] = d
@@ -294,9 +295,6 @@ class Product(ExprWithIntLimits):
         return self.func(self.function.conjugate(), *self.limits)
 
     def _eval_product(self, term, limits):
-        from sympy.concrete.delta import deltaproduct, _has_simple_delta
-        from sympy.concrete.summations import summation
-        from sympy.functions import KroneckerDelta, RisingFactorial
 
         (k, a, n) = limits
 
@@ -308,6 +306,7 @@ class Product(ExprWithIntLimits):
         if a == n:
             return term.subs(k, a)
 
+        from .delta import deltaproduct, _has_simple_delta
         if term.has(KroneckerDelta) and _has_simple_delta(term, limits[0]):
             return deltaproduct(term, limits)
 
@@ -393,7 +392,6 @@ class Product(ExprWithIntLimits):
             return self._eval_product_direct(term, limits)
 
     def _eval_simplify(self, **kwargs):
-        from sympy.simplify.simplify import product_simplify
         rv = product_simplify(self)
         return rv.doit() if kwargs['doit'] else rv
 
@@ -407,7 +405,6 @@ class Product(ExprWithIntLimits):
         return Mul(*[term.subs(k, a + i) for i in range(n - a + 1)])
 
     def _eval_derivative(self, x):
-        from sympy.concrete.summations import Sum
         if isinstance(x, Symbol) and x not in self.free_symbols:
             return S.Zero
         f, limits = self.function, list(self.limits)
@@ -469,8 +466,6 @@ class Product(ExprWithIntLimits):
 
         .. [1] https://en.wikipedia.org/wiki/Infinite_product
         """
-        from sympy.concrete.summations import Sum
-
         sequence_term = self.function
         log_sum = log(sequence_term)
         lim = self.limits
