+++ sympy/concrete/summations.py
@@ -2,9 +2,9 @@ from typing import Tuple as tTuple
 
 from sympy.calculus.singularities import is_decreasing
 from sympy.calculus.util import AccumulationBounds
-from sympy.concrete.expr_with_limits import AddWithLimits
-from sympy.concrete.expr_with_intlimits import ExprWithIntLimits
-from sympy.concrete.gosper import gosper_sum
+from .expr_with_intlimits import ExprWithIntLimits
+from .expr_with_limits import AddWithLimits
+from .gosper import gosper_sum
 from sympy.core.expr import Expr
 from sympy.core.add import Add
 from sympy.core.containers import Tuple
@@ -15,21 +15,31 @@ from sympy.core.relational import Eq
 from sympy.core.singleton import S
 from sympy.core.sorting import ordered
 from sympy.core.symbol import Dummy, Wild, Symbol, symbols
-from sympy.functions.special.zeta_functions import zeta
+from sympy.functions.combinatorial.factorials import factorial
+from sympy.functions.combinatorial.numbers import bernoulli, harmonic
+from sympy.functions.elementary.exponential import log
 from sympy.functions.elementary.piecewise import Piecewise
 from sympy.functions.elementary.trigonometric import cot, csc
+from sympy.functions.special.hyper import hyper
+from sympy.functions.special.tensor_functions import KroneckerDelta
+from sympy.functions.special.zeta_functions import zeta
+from sympy.integrals.integrals import Integral
 from sympy.logic.boolalg import And
-from sympy.polys import apart, together
+from sympy.polys.partfrac import apart
 from sympy.polys.polyerrors import PolynomialError, PolificationFailed
-from sympy.polys.polytools import parallel_poly_from_expr
+from sympy.polys.polytools import parallel_poly_from_expr, Poly, factor
+from sympy.polys.rationaltools import together
 from sympy.series.limitseq import limit_seq
 from sympy.series.order import O
 from sympy.series.residues import residue
-from sympy.sets.sets import FiniteSet
-from sympy.simplify import denom, simplify
+from sympy.sets.sets import FiniteSet, Interval
 from sympy.simplify.combsimp import combsimp
+from sympy.simplify.hyperexpand import hyperexpand
 from sympy.simplify.powsimp import powsimp
-from sympy.solvers import solve
+from sympy.simplify.radsimp import denom, fraction
+from sympy.simplify.simplify import (factor_sum, sum_combine, simplify,
+                                     nsimplify, hypersimp)
+from sympy.solvers.solvers import solve
 from sympy.solvers.solveset import solveset
 from sympy.utilities.iterables import sift
 import itertools
@@ -329,7 +339,6 @@ class Sum(AddWithLimits, ExprWithIntLimits):
         return Sum(f, (k, upper + 1, new_upper)).doit()
 
     def _eval_simplify(self, **kwargs):
-        from sympy.simplify.simplify import factor_sum, sum_combine
 
         # split the function into adds
         terms = Add.make_args(expand(self.function))
@@ -436,9 +445,6 @@ class Sum(AddWithLimits, ExprWithIntLimits):
         Sum.is_absolutely_convergent()
         sympy.concrete.products.Product.is_convergent()
         """
-        from sympy.functions.elementary.exponential import log
-        from sympy.integrals.integrals import Integral
-        from sympy.sets.sets import Interval
         p, q, r = symbols('p q r', cls=Wild)
 
         sym = self.limits[0][0]
@@ -724,9 +730,6 @@ class Sum(AddWithLimits, ExprWithIntLimits):
         With a nonzero eps specified, the summation is ended
         as soon as the remainder term is less than the epsilon.
         """
-        from sympy.functions import bernoulli, factorial
-        from sympy.integrals import Integral
-
         m = int(m)
         n = int(n)
         f = self.function
@@ -997,9 +1000,6 @@ def telescopic(L, R, limits):
 
 
 def eval_sum(f, limits):
-    from sympy.concrete.delta import deltasummation, _has_simple_delta
-    from sympy.functions import KroneckerDelta
-
     (i, a, b) = limits
     if f.is_zero:
         return S.Zero
@@ -1021,6 +1021,7 @@ def eval_sum(f, limits):
             return f.func(*newargs)
 
     if f.has(KroneckerDelta):
+        from .delta import deltasummation, _has_simple_delta
         f = f.replace(
             lambda x: isinstance(x, Sum),
             lambda x: x.factor()
@@ -1106,8 +1107,6 @@ def eval_sum_direct(expr, limits):
 
 
 def eval_sum_symbolic(f, limits):
-    from sympy.functions import harmonic, bernoulli
-
     f_orig = f
     (i, a, b) = limits
     if not f.has(i):
@@ -1247,10 +1246,6 @@ def eval_sum_symbolic(f, limits):
 
 def _eval_sum_hyper(f, i, a):
     """ Returns (res, cond). Sums from a to oo. """
-    from sympy.functions import hyper
-    from sympy.simplify import hyperexpand, hypersimp, fraction
-    from sympy.polys.polytools import Poly, factor
-
     if a != 0:
         return _eval_sum_hyper(f.subs(i, i + a), i, 0)
 
@@ -1264,7 +1259,6 @@ def _eval_sum_hyper(f, i, a):
         return None
 
     if isinstance(hs, Float):
-        from sympy.simplify.simplify import nsimplify
         hs = nsimplify(hs)
 
     numer, denom = fraction(factor(hs))
