+++ sympy/integrals/deltafunctions.py
@@ -1,7 +1,9 @@
-from sympy.core import Mul
-from sympy.functions import DiracDelta, Heaviside
+from sympy.core.mul import Mul
 from sympy.core.singleton import S
 from sympy.core.sorting import default_sort_key
+from sympy.functions import DiracDelta, Heaviside
+from .integrals import Integral, integrate
+from sympy.solvers import solve
 
 
 def change_mul(node, x):
@@ -135,9 +137,6 @@ def deltaintegrate(f, x):
     """
     if not f.has(DiracDelta):
         return None
-
-    from sympy.integrals import Integral, integrate
-    from sympy.solvers import solve
 
     # g(x) = DiracDelta(h(x))
     if f.func == DiracDelta:
