+++ sympy/integrals/risch.py
@@ -40,8 +40,8 @@ from sympy.functions.elementary.hyperbolic import (cosh, coth, sinh,
 from sympy.functions.elementary.piecewise import Piecewise
 from sympy.functions.elementary.trigonometric import (atan, sin, cos,
     tan, acot, cot, asin, acos)
-from sympy.integrals import integrate, Integral
-from sympy.integrals.heurisch import _symbols
+from .integrals import integrate, Integral
+from .heurisch import _symbols
 from sympy.polys.polyerrors import DomainError, PolynomialError
 from sympy.polys.polytools import (real_roots, cancel, Poly, gcd,
     reduced)
@@ -296,6 +296,8 @@ class DifferentialExtension:
         """
         Rewrite exps/pows for better processing.
         """
+        from .prde import is_deriv_k
+
         # Pre-preparsing.
         #################
         # Get all exp arguments, so we can avoid ahead of time doing
@@ -307,8 +309,6 @@ class DifferentialExtension:
         # to a**(Rational*b) before doing anything else.  Note that the
         # _exp_part code can generate terms of this form, so we do need to
         # do this at each pass (or else modify it to not do that).
-
-        from sympy.integrals.prde import is_deriv_k
 
         ratpows = [i for i in self.newf.atoms(Pow).union(self.newf.atoms(exp))
             if (i.base.is_Pow or isinstance(i.base, exp) and i.exp.is_Rational)]
@@ -461,8 +461,7 @@ class DifferentialExtension:
         way around an algebraic extension (e.g., exp(log(x)/2)), it will raise
         NotImplementedError.
         """
-        from sympy.integrals.prde import is_log_deriv_k_t_radical
-
+        from .prde import is_log_deriv_k_t_radical
         new_extension = False
         restart = False
         expargs = [i.exp for i in exps]
@@ -572,8 +571,7 @@ class DifferentialExtension:
         way, so this function does not ever return None or raise
         NotImplementedError.
         """
-        from sympy.integrals.prde import is_deriv_k
-
+        from .prde import is_deriv_k
         new_extension = False
         logargs = [i.args[0] for i in logs]
         for arg in ordered(logargs):
@@ -1389,14 +1387,13 @@ def integrate_primitive_polynomial(p, DE):
     True, or r = p - Dq does not have an elementary integral over k(t) if b is
     False.
     """
-    from sympy.integrals.prde import limited_integrate
-
     Zero = Poly(0, DE.t)
     q = Poly(0, DE.t)
 
     if not p.expr.has(DE.t):
         return (Zero, p, True)
 
+    from .prde import limited_integrate
     while True:
         if not p.expr.has(DE.t):
             return (q, p, True)
@@ -1484,8 +1481,6 @@ def integrate_hyperexponential_polynomial(p, DE, z):
     k[t, 1/t] and a bool b in {True, False} such that p - Dq in k if b is True,
     or p - Dq does not have an elementary integral over k(t) if b is False.
     """
-    from sympy.integrals.rde import rischDE
-
     t1 = DE.t
     dtt = DE.d.exquo(Poly(DE.t, DE.t))
     qa = Poly(0, DE.t)
@@ -1494,6 +1489,8 @@ def integrate_hyperexponential_polynomial(p, DE, z):
 
     if p.is_zero:
         return(qa, qd, b)
+
+    from sympy.integrals.rde import rischDE
 
     with DecrementLevel(DE):
         for i in range(-p.degree(z), p.degree(t1) + 1):
